Title: Update exclusions in wpdeploy.yml
Changes in this commit: Changed the exclusions in the wpdeploy.yml file to exclude files starting with ".git" instead of explicitly listing each file individually.

What is this change for? This change updates the exclusions in the wpdeploy.yml file to make it easier to exclude all files starting with ".git" instead of listing each file individually.

What security risks exist? There are no security risks associated with this change.